Add expression display
Add a UILabel that displays the operations performed that arrived
at the result shown in the calculator display. This solution uses
a second enum in CalculatorBrain which maybe isn't a good choice,
but it is functional according to the programming assignment. (shurg)

checkpoint: Label added w/ IBOutlet

Things work except for 7 + = would show '7 + 7 =' (14 in the display)

Working